requirements.txt
# ═══════════════════════════════════════════
# 퀀트 트레이딩 시스템 — 의존성 패키지
# ═══════════════════════════════════════════

# 데이터 수집
pykrx>=1.0.45
requests>=2.31.0

# 데이터 분석
pandas>=2.1.0
numpy>=1.24.0

# 기술적 지표
ta>=0.11.0

# 웹 대시보드 (FastAPI + Streamlit)
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.3
streamlit>=1.31.0

# 텔레그램 알림
python-telegram-bot>=20.7

# 스케줄링
schedule>=1.2.1
apscheduler>=3.10.4

# 유틸리티
python-dotenv>=1.0.0
aiohttp>=3.9.0
httpx>=0.27.0
